LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src/script-src-strict_dynamic_in_img-src-expected.txt

CONSOLE MESSAGE: The source list for Content Security Policy directive 'img-src' contains an invalid source: ''strict-dynamic''. It will be ignored.
`strict-dynamic` does not drop allowed source expressions in `img-src`.


PASS `strict-dynamic` does not drop allowed source expressions in `img-src`.
